# Adding an `OnlyKeysVisitor` to reduce the config size when serializing