--- a/src/Sulu/Bundle/AdminBundle/Resources/config/services.xml
+++ b/src/Sulu/Bundle/AdminBundle/Resources/config/services.xml
@@ -68,6 +68,13 @@
         </service>
 
         <service
+            id="sulu_admin.only_keys_metadata_visitor"
+            class="Sulu\Bundle\AdminBundle\Metadata\FormMetadata\OnlyKeysVisistor"
+        >
+            <tag name="sulu_admin.typed_form_metadata_visitor" />
+        </service>
+
+        <service
             id="sulu_admin.tag_filter_typed_form_metadata_visitor"
             class="Sulu\Bundle\AdminBundle\Metadata\FormMetadata\TagFilterTypedFormMetadataVisitor"
         >